Sistema de marca páginas terminado

+++ app/Http/Controllers/BookmarkController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Book;
+use App\Bookmark;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -18,7 +19,12 @@ class BookmarkController extends Controller
     public function index(Request $request)
     {
         $bookID = $request->input('book');
-        $bookmarks = Book::find($bookID)->bookmarks()->get();
+        $bookmarks = [];
+
+        if(Auth::user()->books->contains($bookID))
+        {
+            $bookmarks = Book::find($bookID)->bookmarks()->get();
+        }
 
         return response()->json([
             'bookmarks' => $bookmarks
@@ -27,12 +33,50 @@ class BookmarkController extends Controller
 
     public function add(Request $request)
     {
+        $status = false;
+        $bookID = $request->input('book');
 
+        if(Auth::user()->books->contains($bookID))
+        {
+            $bookmark = new Bookmark();
+            $bookmark->book_id = $request->input('book');
+            $bookmark->page = $request->input('page');
+            $bookmark->comment = $request->input('comment');
+            $bookmark->save();
+
+            $status = true;
+        }
+
+        return response()->json([
+            'status' => $status
+        ]);
     }
 
     public function update(Request $request)
     {
+        $status = false;
 
+        $bookID = $request->input('book');
+        $bookmarkID = $request->input('bookmark');
+        $bookmarks = Book::find($bookID)->bookmarks()->get();
+
+        if(Auth::user()->books->contains($bookID) && $bookmarks->contains($bookmarkID))
+        {
+            $bookmark = $bookmarks->find($bookmarkID);
+
+            if($bookmark)
+            {
+                $bookmark->comment = $request->input('comment');
+                $bookmark->save();
+
+                $status = true;
+            }
+        }
+
+        return response()->json([
+            'status' => $status,
+            'id' => $bookmarkID
+        ]);
     }
 
     public function delete(Request $request)
@@ -43,7 +87,7 @@ class BookmarkController extends Controller
         $bookmarkID = $request->input('bookmark');
         $bookmarks = Book::find($bookID)->bookmarks()->get();
 
-        if($bookmarks->contains($bookmarkID))
+        if(Auth::user()->books->contains($bookID) && $bookmarks->contains($bookmarkID))
         {
             if($bookmarks->find($bookmarkID)->delete())
             {
